chore: merge BaseLoader.ts into loaders.ts
